+++ taglib/toolkit/tiostream.h
@@ -87,7 +87,7 @@ namespace TagLib {
     /*!
      * Reads a block of size \a length at the current get pointer.
      */
-    virtual ByteVector readBlock(uint length) = 0;
+    virtual ByteVector readBlock(size_t length) = 0;
 
     /*!
      * Attempts to write the block \a data at the current get pointer.  If the
@@ -107,7 +107,7 @@ namespace TagLib {
      * \note This method is slow since it requires rewriting all of the file
      * after the insertion point.
      */
-    virtual void insert(const ByteVector &data, offset_t start = 0, uint replace = 0) = 0;
+    virtual void insert(const ByteVector &data, offset_t start = 0, size_t replace = 0) = 0;
 
     /*!
      * Removes a block of the file starting a \a start and continuing for
@@ -116,7 +116,7 @@ namespace TagLib {
      * \note This method is slow since it involves rewriting all of the file
      * after the removed portion.
      */
-    virtual void removeBlock(offset_t start = 0, uint length = 0) = 0;
+    virtual void removeBlock(offset_t start = 0, size_t length = 0) = 0;
 
     /*!
      * Returns true if the file is read only (or if the file can not be opened).
